Add party index CSS positioning
