+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/alex/partyr/frontend/src/components/livechat/livechat.jsx":"1","/home/alex/partyr/frontend/src/components/party/index/party_index_map.jsx":"2","/home/alex/partyr/frontend/src/components/party/index/party_index_item.jsx":"3"},{"size":3934,"mtime":1607016954170,"results":"4","hashOfConfig":"5"},{"size":2076,"mtime":1607021780400,"results":"6","hashOfConfig":"5"},{"size":1021,"mtime":1607024690810,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"3kb444",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/alex/partyr/frontend/src/components/livechat/livechat.jsx",[],"/home/alex/partyr/frontend/src/components/party/index/party_index_map.jsx",[],"/home/alex/partyr/frontend/src/components/party/index/party_index_item.jsx",[]]
+[{"/home/alex/partyr/frontend/src/components/livechat/livechat.jsx":"1","/home/alex/partyr/frontend/src/components/party/index/party_index_map.jsx":"2","/home/alex/partyr/frontend/src/components/party/index/party_index_item.jsx":"3","/home/alex/partyr/frontend/src/components/party/index/party_index.jsx":"4"},{"size":3917,"mtime":1607024992420,"results":"5","hashOfConfig":"6"},{"size":2381,"mtime":1607029161430,"results":"7","hashOfConfig":"6"},{"size":2945,"mtime":1607029747430,"results":"8","hashOfConfig":"6"},{"size":956,"mtime":1607029404420,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"3kb444",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/alex/partyr/frontend/src/components/livechat/livechat.jsx",[],"/home/alex/partyr/frontend/src/components/party/index/party_index_map.jsx",[],"/home/alex/partyr/frontend/src/components/party/index/party_index_item.jsx",[],"/home/alex/partyr/frontend/src/components/party/index/party_index.jsx",["18"],{"ruleId":"19","severity":1,"message":"20","line":32,"column":5,"nodeType":"21","messageId":"22","endLine":32,"endColumn":8},"no-unused-vars","'map' is assigned a value but never used.","Identifier","unusedVar"]
