+++ frontend/src/components/livechat/livechat.jsx
@@ -69,7 +69,6 @@ class Party extends React.Component {
 
   componentDidUpdate(prevProps) {
     if (prevProps.party !== this.props.party && !Array.isArray(prevProps.party)) {
-      debugger;
       this.socket.emit('join', {
         partyId: this.props.party._id
       },
