+++ frontend/src/components/party/index/party_index.css
@@ -1,0 +1,48 @@
+.party-index-container {
+  border: 1px solid red;
+  margin-top: 50px;
+}
+
+.party-index {
+  border: 1px solid blue;
+  width: 400px;
+  height: 500px;
+  margin: 0 auto;
+}
+
+.party-index-item {
+  width: 400px;
+  height: 500px;
+  background-color: white;
+  border: 1px solid orange;
+  border-radius: 15px;
+  margin: 0 auto 10px auto;
+  position: absolute;
+  /* left: calc(100% / 6 + 200px); */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.party-index-map-container {
+  display: block;
+  width: 350px;
+  height: 250px;
+  /* border: 1px solid purple; */
+  margin: 20px auto;
+}
+
+.map-container {
+  border: 1px solid green;
+}
+
+.party-date, .part-time {
+  display: flex;
+  flex-direction: row;
+}
+
+.party-title {
+  font-size: 24px;
+  font-weight: 700;
+  margin-top: 15px;
+}
